chancges after PR comments
