--- pkg/module/source/source.go
@@ -1,15 +1,0 @@
-package source
-
-import (
-	ocm "github.com/open-component-model/ocm/pkg/contexts/ocm/compdesc"
-	"github.com/open-component-model/ocm/pkg/contexts/ocm/cpi"
-)
-
-const (
-	ocmIdentityName = "module-sources"
-	ocmVersion      = "v1"
-)
-
-type Source interface {
-	FetchSource(ctx cpi.Context, path, repo, version string) (*ocm.Source, error)
-}
